Refactor forecast.py
Modify forecast.py, clean up functions and moving a nested
function to make the code easier to follow. Also clean up
and reduce code where posible.

The goal is to make the functions more portable and prepare
to move some code into classes to make it even more portable
in the future.