Add file 'wind_forecast.py'
'wind_forecast.py' uses Matplotlib to plot wind direction and velocity
in both a polar barchart for direction / windspeed and a stemplot for
windspeed / time. It can be passed latitude and longitude coordinates,
or if none are passed it will return data for Minneapolis, MN.
The data returned is for a 24 hour period beginning with the current
time.